pr-157546: Split TransportStatelessPrimaryRelocationAction
