+++ x-pack/plugin/stateless/src/internalClusterTest/java/org/elasticsearch/xpack/stateless/recovery/IndexingShardRelocationIT.java
@@ -119,7 +119,7 @@ import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertNoFa
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertResponse;
 import static org.elasticsearch.xpack.stateless.commits.HollowShardsService.STATELESS_HOLLOW_INDEX_SHARDS_ENABLED;
 import static org.elasticsearch.xpack.stateless.objectstore.ObjectStoreTestUtils.getObjectStoreMockRepository;
-import static org.elasticsearch.xpack.stateless.recovery.TransportStatelessPrimaryRelocationAction.MAX_SLOW_OPERATION_THREAD_DUMPS;
+import static org.elasticsearch.xpack.stateless.recovery.SlowRelocationLogger.MAX_SLOW_OPERATION_THREAD_DUMPS;
 import static org.elasticsearch.xpack.stateless.recovery.TransportStatelessPrimaryRelocationAction.PRIMARY_CONTEXT_HANDOFF_ACTION_NAME;
 import static org.elasticsearch.xpack.stateless.recovery.TransportStatelessPrimaryRelocationAction.SLOW_RELOCATION_THRESHOLD_SETTING;
 import static org.elasticsearch.xpack.stateless.recovery.TransportStatelessPrimaryRelocationAction.START_RELOCATION_ACTION_NAME;
@@ -804,7 +804,7 @@ public class IndexingShardRelocationIT extends AbstractStatelessPluginIntegTestC
 
     @TestLogging(
         reason = "verifying INFO logging of repeated hot threads dumps",
-        value = "org.elasticsearch.xpack.stateless.recovery.TransportStatelessPrimaryRelocationAction:INFO"
+        value = "org.elasticsearch.xpack.stateless.recovery.SlowRelocationLogger:INFO"
     )
     public void testSlowRelocationLogsRepeatedHotThreadsDumps() throws Exception {
         final var nodeSettings = Settings.builder().put(SLOW_RELOCATION_THRESHOLD_SETTING.getKey(), TimeValue.timeValueMillis(100)).build();
@@ -821,12 +821,12 @@ public class IndexingShardRelocationIT extends AbstractStatelessPluginIntegTestC
             .getShard(0);
         final var permitFuture = new PlainActionFuture<Releasable>();
         indexShard.acquirePrimaryOperationPermit(permitFuture, EsExecutors.DIRECT_EXECUTOR_SERVICE);
-        try (Releasable permit = safeGet(permitFuture); var mockLog = MockLog.capture(TransportStatelessPrimaryRelocationAction.class)) {
+        try (Releasable permit = safeGet(permitFuture); var mockLog = MockLog.capture(SlowRelocationLogger.class)) {
             for (int sample = 1; sample <= MAX_SLOW_OPERATION_THREAD_DUMPS; sample++) {
                 mockLog.addExpectation(
                     new MockLog.SeenEventExpectation(
                         "hot threads dump " + sample,
-                        TransportStatelessPrimaryRelocationAction.class.getCanonicalName(),
+                        SlowRelocationLogger.class.getCanonicalName(),
                         Level.INFO,
                         "* recovery [*]: flush and acquire permits #" + sample + " with [*] operations holding permits*"
                     )
