+++ x-pack/plugin/stateless/src/test/java/org/elasticsearch/xpack/stateless/snapshots/StatelessSnapshotResiliencyTests.java
@@ -144,6 +144,8 @@ import org.elasticsearch.xpack.stateless.recovery.PitRelocationMetrics;
 import org.elasticsearch.xpack.stateless.recovery.RecoveryCommitRegistrationHandler;
 import org.elasticsearch.xpack.stateless.recovery.RemoveRefreshClusterBlockService;
 import org.elasticsearch.xpack.stateless.recovery.StatelessIndexNodeRecoveryListener;
+import org.elasticsearch.xpack.stateless.recovery.StatelessPrimaryRelocationSourceService;
+import org.elasticsearch.xpack.stateless.recovery.StatelessPrimaryRelocationTargetService;
 import org.elasticsearch.xpack.stateless.recovery.StatelessSearchNodeRecoveryListener;
 import org.elasticsearch.xpack.stateless.recovery.TransportRegisterCommitForRecoveryAction;
 import org.elasticsearch.xpack.stateless.recovery.TransportSendRecoveryCommitRegistrationAction;
@@ -506,15 +508,27 @@ public class StatelessSnapshotResiliencyTests extends SnapshotResiliencyTests {
                     StatelessPrimaryRelocationAction.TYPE,
                     new TransportStatelessPrimaryRelocationAction(
                         transportService(),
-                        clusterService(),
                         actionFilters,
                         indicesService,
                         new CompositeRecoverySchedulingListener(),
+                        new StatelessPrimaryRelocationSourceService(
+                            clusterService(),
+                            transportService().getThreadPool(),
+                            indicesService,
+                            testStatelessPlugin.hollowShardsService,
+                            new StatelessCommitServiceProvider(testStatelessPlugin.statelessCommitService),
+                            mock(IndexShardCacheWarmer.class),
+                            HollowShardsMetrics.NOOP
+                        ),
+                        new StatelessPrimaryRelocationTargetService(
+                            clusterService(),
+                            transportService().getThreadPool(),
+                            indicesService,
+                            new StatelessCommitServiceProvider(testStatelessPlugin.statelessCommitService),
+                            mock(IndexShardCacheWarmer.class),
+                            new StatelessPrimaryRelocationMetricsCollectorProvider(StatelessPrimaryRelocationMetricsCollector.NOOP)
+                        ),
                         peerRecoveryTargetService,
-                        new StatelessCommitServiceProvider(testStatelessPlugin.statelessCommitService),
-                        mock(IndexShardCacheWarmer.class),
-                        testStatelessPlugin.hollowShardsService,
-                        HollowShardsMetrics.NOOP,
                         new StatelessPrimaryRelocationMetricsCollectorProvider(StatelessPrimaryRelocationMetricsCollector.NOOP)
                     ),
                     StatelessUnpromotableRelocationAction.TYPE,
